BLD: upgraded integrationTests setup
