+++ .gitignore
@@ -74,5 +74,4 @@ enigmacontract.txt
 enigmatokencontract.txt
 
 # integrationTests
-enigma-js/test/integrationTests/*.spec.js
 enigma-js/test/integrationTests/testList.txt
