+++ enigma-js/test/integrationTests/02_deploy_calculator.spec.js
@@ -6,7 +6,8 @@ import Web3 from 'web3';
 import Enigma from '../../src/Enigma';
 import utils from '../../src/enigma-utils';
 import * as eeConstants from '../../src/emitterConstants';
-import {EnigmaContract, EnigmaTokenContract} from './contractLoader';
+import {EnigmaContract, EnigmaTokenContract, EnigmaContractAddress,
+  EnigmaTokenContractAddress, proxyAddress, ethNodeAddr} from './contractLoader';
 import * as constants from './testConstants';
 
 
@@ -18,16 +19,17 @@ describe('Enigma tests', () => {
   let accounts;
   let web3;
   let enigma;
+  let epochSize;
   it('initializes', () => {
-    const provider = new Web3.providers.HttpProvider('http://localhost:9545');
+    const provider = new Web3.providers.HttpProvider(ethNodeAddr);
     web3 = new Web3(provider);
     return web3.eth.getAccounts().then((result) => {
       accounts = result;
       enigma = new Enigma(
         web3,
-        EnigmaContract.networks['4447'].address,
-        EnigmaTokenContract.networks['4447'].address,
-        'http://localhost:3346',
+        EnigmaContractAddress,
+        EnigmaTokenContractAddress,
+        proxyAddress,
         {
           gas: 4712388,
           gasPrice: 100000000000,
@@ -35,6 +37,7 @@ describe('Enigma tests', () => {
         },
       );
       enigma.admin();
+      enigma.setTaskKeyPair('cupcake');
       expect(Enigma.version()).toEqual('0.0.1');
     });
   });
@@ -43,20 +46,14 @@ describe('Enigma tests', () => {
   let task;
   const homedir = os.homedir();
 
-  it('should generate and save key/pair', () => {
-    enigma.setTaskKeyPair('cupcake');
-  });
-
   it('should deploy secret contract', async () => {
-    let scTaskFn = `construct(address)`;
-    let scTaskArgs = [
-      ['0x0000000000000000000000000000000000000102', 'address'],
-    ];
-    let scTaskGasLimit = 4000000;
+    let scTaskFn = 'construct()';
+    let scTaskArgs = '';
+    let scTaskGasLimit = 1000000;
     let scTaskGasPx = utils.toGrains(1);
     let preCode;
     try {
-      preCode = fs.readFileSync(path.resolve(__dirname,'secretContracts/voting.wasm'));
+      preCode = fs.readFileSync(path.resolve(__dirname,'secretContracts/calculator.wasm'));
     } catch(e) {
       console.log('Error:', e.stack);
     }
@@ -66,7 +63,7 @@ describe('Enigma tests', () => {
         .on(eeConstants.ERROR, (error) => reject(error));
     });
 
-    fs.writeFile(path.join(homedir, '.enigma', 'addr-voting-wrongeth.txt'), scTask.scAddr, 'utf8', function(err) {
+    fs.writeFile(path.join(homedir, '.enigma', 'addr-calculator.txt'), scTask.scAddr, 'utf8', function(err) {
       if(err) {
         return console.log(err);
       }
@@ -78,7 +75,7 @@ describe('Enigma tests', () => {
       await sleep(1000);
       scTask = await enigma.getTaskRecordStatus(scTask);
       process.stdout.write('Waiting. Current Task Status is '+scTask.ethStatus+'\r');
-    } while (scTask.ethStatus !== 2);
+    } while (scTask.ethStatus != 2);
     expect(scTask.ethStatus).toEqual(2);
     process.stdout.write('Completed. Final Task Status is '+scTask.ethStatus+'\n');
   }, constants.TIMEOUT_DEPLOY);
@@ -90,7 +87,7 @@ describe('Enigma tests', () => {
 
   it('should get deployed contract bytecode hash', async () => {
     const result = await enigma.admin.getCodeHash(scTask.scAddr);
-    expect(result).toBeTruthy();
+    expect(result).toBeTruthy;
     console.log('Deployed contract bytecode hash is: '+result);
   });
 });
