+++ enigma-js/test/integrationTests/03_deploy_fail_parameters.spec.js
@@ -6,7 +6,8 @@ import Web3 from 'web3';
 import Enigma from '../../src/Enigma';
 import utils from '../../src/enigma-utils';
 import * as eeConstants from '../../src/emitterConstants';
-import {EnigmaContract, EnigmaTokenContract} from './contractLoader';
+import {EnigmaContract, EnigmaTokenContract, EnigmaContractAddress, EnigmaTokenContractAddress,
+  proxyAddress, ethNodeAddr} from './contractLoader';
 import * as constants from './testConstants';
 
 
@@ -20,15 +21,15 @@ describe('Enigma tests', () => {
   let enigma;
   let epochSize;
   it('initializes', () => {
-    const provider = new Web3.providers.HttpProvider('http://localhost:9545');
+    const provider = new Web3.providers.HttpProvider(ethNodeAddr);
     web3 = new Web3(provider);
     return web3.eth.getAccounts().then((result) => {
       accounts = result;
       enigma = new Enigma(
         web3,
-        EnigmaContract.networks['4447'].address,
-        EnigmaTokenContract.networks['4447'].address,
-        'http://localhost:3346',
+        EnigmaContractAddress,
+        EnigmaTokenContractAddress,
+        proxyAddress,
         {
           gas: 4712388,
           gasPrice: 100000000000,
@@ -36,36 +37,29 @@ describe('Enigma tests', () => {
         },
       );
       enigma.admin();
+      enigma.setTaskKeyPair('cupcake');
       expect(Enigma.version()).toEqual('0.0.1');
     });
   });
 
-  it('should generate and save key/pair', () => {
-    enigma.setTaskKeyPair('cupcake');
-  });
-
-  // The following tests below currently don't fail.
-  // This issue is tracked here: https://github.com/enigmampc/enigma-core/issues/111
-
   let scTask2;
-  it('should deploy secret contract with wrong constructor', async () => {
-    let scTaskFn = 'not a valid construct()';
-    let scTaskArgs = '';
-    let scTaskGasLimit = 100;
+  it('should deploy secret contract', async () => {
+    let scTaskFn = 'construct()';
+    let scTaskArgs =  '';   // Wrong parameters, expecting ETH address
+    let scTaskGasLimit = 1000000;
     let scTaskGasPx = utils.toGrains(1);
     let preCode;
     try {
-      preCode = fs.readFileSync(path.resolve(__dirname,'secretContracts/calculator.wasm'));
+      preCode = fs.readFileSync(path.resolve(__dirname,'secretContracts/voting.wasm'));
     } catch(e) {
       console.log('Error:', e.stack);
     }
-
     scTask2 = await new Promise((resolve, reject) => {
       enigma.deploySecretContract(scTaskFn, scTaskArgs, scTaskGasLimit, scTaskGasPx, accounts[0], preCode)
         .on(eeConstants.DEPLOY_SECRET_CONTRACT_RESULT, (receipt) => resolve(receipt))
         .on(eeConstants.ERROR, (error) => reject(error));
     });
-  }, constants.TIMEOUT_FAILDEPLOY);
+  }, constants.TIMEOUT_DEPLOY);
 
   it('should get the failed receipt', async () => {
     do {
@@ -73,9 +67,9 @@ describe('Enigma tests', () => {
       scTask2 = await enigma.getTaskRecordStatus(scTask2);
       process.stdout.write('Waiting. Current Task Status is '+scTask2.ethStatus+'\r');
     } while (scTask2.ethStatus != 3);
-    expect(scTask2.ethStatus).toEqual(3);   // <- it succeeds and returns 2 instead for a successful deployment
+    expect(scTask2.ethStatus).toEqual(3);
     process.stdout.write('Completed. Final Task Status is '+scTask2.ethStatus+'\n');
-  }, constants.TIMEOUT_FAILDEPLOY);
+  }, constants.TIMEOUT_DEPLOY);
 
   it('should fail to verify deployed contract', async () => {
     const result = await enigma.admin.isDeployed(scTask2.scAddr);
@@ -86,6 +80,5 @@ describe('Enigma tests', () => {
     const result = await enigma.admin.getCodeHash(scTask2.scAddr);
     expect(result).toBeFalsy;
   });
-
 
 });
