+++ enigma-js/test/integrationTests/10_execute_flipcoin.spec.js
@@ -6,7 +6,8 @@ import Web3 from 'web3';
 import Enigma from '../../src/Enigma';
 import utils from '../../src/enigma-utils';
 import * as eeConstants from '../../src/emitterConstants';
-import {EnigmaContract, EnigmaTokenContract} from './contractLoader';
+import {EnigmaContract, EnigmaTokenContract, EnigmaContractAddress, EnigmaTokenContractAddress,
+  proxyAddress, ethNodeAddr} from './contractLoader';
 import * as constants from './testConstants';
 
 
@@ -20,15 +21,15 @@ describe('Enigma tests', () => {
   let enigma;
   let epochSize;
   it('initializes', () => {
-    const provider = new Web3.providers.HttpProvider('http://localhost:9545');
+    const provider = new Web3.providers.HttpProvider(ethNodeAddr);
     web3 = new Web3(provider);
     return web3.eth.getAccounts().then((result) => {
       accounts = result;
       enigma = new Enigma(
         web3,
-        EnigmaContract.networks['4447'].address,
-        EnigmaTokenContract.networks['4447'].address,
-        'http://localhost:3346',
+        EnigmaContractAddress,
+        EnigmaTokenContractAddress,
+        proxyAddress,
         {
           gas: 4712388,
           gasPrice: 100000000000,
@@ -36,53 +37,55 @@ describe('Enigma tests', () => {
         },
       );
       enigma.admin();
+      enigma.setTaskKeyPair('cupcake');
       expect(Enigma.version()).toEqual('0.0.1');
     });
   });
 
-  it('should generate and save key/pair', () => {
-    enigma.setTaskKeyPair('cupcake');
-  });
+  const homedir = os.homedir();
 
+  const flipcoinAddr = fs.readFileSync(path.join(homedir, '.enigma', 'addr-flipcoin.txt'), 'utf-8');
   let task;
   it('should execute compute task', async () => {
-    let nonExistentAddr = '0x1234567890123456789012345678901234567890123456789012345678901234'
-    let taskFn = 'function()';
-    let taskArgs = '';
-    let taskGasLimit = 1000;
+    let taskFn = 'flip()';
+    let taskArgs = [];
+    let taskGasLimit = 100000;
     let taskGasPx = utils.toGrains(1);
     task = await new Promise((resolve, reject) => {
-      enigma.computeTask(taskFn, taskArgs, taskGasLimit, taskGasPx, accounts[0], nonExistentAddr)
+      enigma.computeTask(taskFn, taskArgs, taskGasLimit, taskGasPx, accounts[0], flipcoinAddr)
         .on(eeConstants.SEND_TASK_INPUT_RESULT, (result) => resolve(result))
         .on(eeConstants.ERROR, (error) => reject(error));
     });
-  });
+  }, constants.TIMEOUT_COMPUTE);
 
   it('should get the pending task', async () => {
     task = await enigma.getTaskRecordStatus(task);
     expect(task.ethStatus).toEqual(1);
   });
 
-  it('should get the failed task receipt', async () => {
+  it('should get the confirmed task', async () => {
     do {
       await sleep(1000);
       task = await enigma.getTaskRecordStatus(task);
       process.stdout.write('Waiting. Current Task Status is '+task.ethStatus+'\r');
-    } while (task.ethStatus != 3);
-    expect(task.ethStatus).toEqual(3);
+    } while (task.ethStatus != 2);
+    expect(task.ethStatus).toEqual(2);
     process.stdout.write('Completed. Final Task Status is '+task.ethStatus+'\n');
   }, constants.TIMEOUT_COMPUTE);
 
-  xit('should get the failed result', async () => {
+  it('should get and validate the result', async () => {
     task = await new Promise((resolve, reject) => {
-        enigma.getTaskResult(task)
-          .on(eeConstants.GET_TASK_RESULT_RESULT, (result) => resolve(result))
-          .on(eeConstants.ERROR, (error) => reject(error));
+      enigma.getTaskResult(task)
+        .on(eeConstants.GET_TASK_RESULT_RESULT, (result) => resolve(result))
+        .on(eeConstants.ERROR, (error) => reject(error));
     });
-    expect(task.engStatus).toEqual('FAILED');
+    expect(task.engStatus).toEqual('SUCCESS');
     expect(task.encryptedAbiEncodedOutputs).toBeTruthy();
     expect(task.usedGas).toBeTruthy();
     expect(task.workerTaskSig).toBeTruthy();
+    task = await enigma.decryptTaskResult(task);
+    const result = parseInt(task.decryptedOutput, 16);
+    expect(result === 1 || result === 0).toBe(true);
   });
 
 });
