+++ enigma-js/test/integrationTests/10_execute_millionaire.spec.js
@@ -6,9 +6,9 @@ import Web3 from 'web3';
 import Enigma from '../../src/Enigma';
 import utils from '../../src/enigma-utils';
 import * as eeConstants from '../../src/emitterConstants';
-import {EnigmaContract, EnigmaTokenContract} from './contractLoader';
+import {EnigmaContract, EnigmaTokenContract, SampleContract, EnigmaContractAddress, EnigmaTokenContractAddress,
+  proxyAddress, ethNodeAddr} from './contractLoader'
 import * as constants from './testConstants';
-
 
 
 function sleep(ms) {
@@ -21,15 +21,15 @@ describe('Enigma tests', () => {
   let enigma;
   let epochSize;
   it('initializes', () => {
-    const provider = new Web3.providers.HttpProvider('http://localhost:9545');
+    const provider = new Web3.providers.HttpProvider(ethNodeAddr);
     web3 = new Web3(provider);
     return web3.eth.getAccounts().then((result) => {
       accounts = result;
       enigma = new Enigma(
         web3,
-        EnigmaContract.networks['4447'].address,
-        EnigmaTokenContract.networks['4447'].address,
-        'http://localhost:3346',
+        EnigmaContractAddress,
+        EnigmaTokenContractAddress,
+        proxyAddress,
         {
           gas: 4712388,
           gasPrice: 100000000000,
@@ -37,116 +37,124 @@ describe('Enigma tests', () => {
         },
       );
       enigma.admin();
+      enigma.setTaskKeyPair('cupcake');
       expect(Enigma.version()).toEqual('0.0.1');
     });
   });
 
   const homedir = os.homedir();
-  const revertAddr = fs.readFileSync(path.join(homedir, '.enigma', 'addr-revert.txt'), 'utf-8');
 
-  it('should generate and save key/pair', () => {
-    enigma.setTaskKeyPair('cupcake');
-  });
+  const millionaireAddr = fs.readFileSync(path.join(homedir, '.enigma', 'addr-millionaire.txt'), 'utf-8');
 
   let task1;
-  it('should read the state and validate initial value', async () => {
-    let taskFn = 'get_last_sum()';
-    let taskArgs = [];
-    let taskGasLimit = 100000;
+  const addr1 = '0x0000000000000000000000000000000000000000000000000000000000000001';
+  const addr2 = '0x0000000000000000000000000000000000000000000000000000000000000002';
+  it('should execute compute task', async () => {
+    let taskFn = 'add_millionaire(bytes32,uint256)';
+    let taskArgs = [
+        [addr1, 'bytes32'],
+        [1000000, 'uint256'],
+      ];
+    let taskGasLimit = 1000000;
     let taskGasPx = utils.toGrains(1);
     task1 = await new Promise((resolve, reject) => {
-      enigma.computeTask(taskFn, taskArgs, taskGasLimit, taskGasPx, accounts[0], revertAddr)
+      enigma.computeTask(taskFn, taskArgs, taskGasLimit, taskGasPx, accounts[0], millionaireAddr)
         .on(eeConstants.SEND_TASK_INPUT_RESULT, (result) => resolve(result))
         .on(eeConstants.ERROR, (error) => reject(error));
     });
-  }, constants.TIMEOUT_COMPUTE);    
+  }, constants.TIMEOUT_COMPUTE);
+
+  it('should get the pending task', async () => {
+    task1 = await enigma.getTaskRecordStatus(task1);
+    expect(task1.ethStatus).toEqual(eeConstants.ETH_STATUS_CREATED);
+  });
 
   it('should get the confirmed task', async () => {
     do {
       await sleep(1000);
       task1 = await enigma.getTaskRecordStatus(task1);
       process.stdout.write('Waiting. Current Task Status is '+task1.ethStatus+'\r');
-    } while (task1.ethStatus != 2);
-    expect(task1.ethStatus).toEqual(2);
+    } while (task1.ethStatus != eeConstants.ETH_STATUS_VERIFIED);
+    expect(task1.ethStatus).toEqual(eeConstants.ETH_STATUS_VERIFIED);
     process.stdout.write('Completed. Final Task Status is '+task1.ethStatus+'\n');
   }, constants.TIMEOUT_COMPUTE);
 
-  it('should get the result and verify the computation is correct', async () => {
-    task1 = await new Promise((resolve, reject) => {
-      enigma.getTaskResult(task1)
-        .on(eeConstants.GET_TASK_RESULT_RESULT, (result) => resolve(result))
-        .on(eeConstants.ERROR, (error) => reject(error));
-    });
-    expect(task1.engStatus).toEqual('SUCCESS');
-    expect(task1.encryptedAbiEncodedOutputs).toBeTruthy();
-    task1 = await enigma.decryptTaskResult(task1);
-    expect(task1.usedGas).toBeTruthy();
-    expect(task1.workerTaskSig).toBeTruthy();
-    expect(parseInt(task1.decryptedOutput, 16)).toEqual(12);
-  });
-
   let task2;
-  it('should execute sum_and_call, and verify output of computation', async () => {
-    let taskFn = 'sum_and_call(uint256,uint256,address)';
+  it('should execute compute task', async () => {
+    let taskFn = 'add_millionaire(bytes32,uint256)';
     let taskArgs = [
-      [2, 'uint256'],
-      [3, 'uint256'],
-      ["0x9b1f7F645351AF3631a656421eD2e40f2802E6c0", 'address']
-    ];
-    let taskGasLimit = 100000;
+        [addr2, 'bytes32'],
+        [2000000, 'uint256'],
+      ];
+    let taskGasLimit = 1000000;
     let taskGasPx = utils.toGrains(1);
     task2 = await new Promise((resolve, reject) => {
-      enigma.computeTask(taskFn, taskArgs, taskGasLimit, taskGasPx, accounts[0], revertAddr)
+      enigma.computeTask(taskFn, taskArgs, taskGasLimit, taskGasPx, accounts[0], millionaireAddr)
         .on(eeConstants.SEND_TASK_INPUT_RESULT, (result) => resolve(result))
         .on(eeConstants.ERROR, (error) => reject(error));
     });
   }, constants.TIMEOUT_COMPUTE);
 
-  it('should get the confirmed task failure', async () => {
+  it('should get the pending task', async () => {
+    task2 = await enigma.getTaskRecordStatus(task2);
+    expect(task2.ethStatus).toEqual(eeConstants.ETH_STATUS_CREATED);
+  });
+
+  it('should get the confirmed task', async () => {
     do {
       await sleep(1000);
       task2 = await enigma.getTaskRecordStatus(task2);
       process.stdout.write('Waiting. Current Task Status is '+task2.ethStatus+'\r');
-    } while (task2.ethStatus != 4);
-    expect(task2.ethStatus).toEqual(4);
+    } while (task2.ethStatus != eeConstants.ETH_STATUS_VERIFIED);
+    expect(task2.ethStatus).toEqual(eeConstants.ETH_STATUS_VERIFIED);
     process.stdout.write('Completed. Final Task Status is '+task2.ethStatus+'\n');
   }, constants.TIMEOUT_COMPUTE);
 
   let task3;
-  it('should read the state again, and validate the value is still the initial value, despite the write_state!', async () => {
-    let taskFn = 'get_last_sum()';
+  it('should execute compute task', async () => {
+    let taskFn = 'compute_richest()';
     let taskArgs = [];
-    let taskGasLimit = 100000;
+    let taskGasLimit = 1000000;
     let taskGasPx = utils.toGrains(1);
     task3 = await new Promise((resolve, reject) => {
-      enigma.computeTask(taskFn, taskArgs, taskGasLimit, taskGasPx, accounts[0], revertAddr)
+      enigma.computeTask(taskFn, taskArgs, taskGasLimit, taskGasPx, accounts[0], millionaireAddr)
         .on(eeConstants.SEND_TASK_INPUT_RESULT, (result) => resolve(result))
         .on(eeConstants.ERROR, (error) => reject(error));
     });
-  }, constants.TIMEOUT_COMPUTE);    
+  }, constants.TIMEOUT_COMPUTE);
+
+  it('should get the pending task', async () => {
+    task3 = await enigma.getTaskRecordStatus(task3);
+    expect(task3.ethStatus).toEqual(eeConstants.ETH_STATUS_CREATED);
+  });
 
   it('should get the confirmed task', async () => {
     do {
       await sleep(1000);
       task3 = await enigma.getTaskRecordStatus(task3);
       process.stdout.write('Waiting. Current Task Status is '+task3.ethStatus+'\r');
-    } while (task3.ethStatus != 2);
-    expect(task3.ethStatus).toEqual(2);
+    } while (task3.ethStatus != eeConstants.ETH_STATUS_VERIFIED);
+    expect(task3.ethStatus).toEqual(eeConstants.ETH_STATUS_VERIFIED);
     process.stdout.write('Completed. Final Task Status is '+task3.ethStatus+'\n');
   }, constants.TIMEOUT_COMPUTE);
 
-  it('should get the result and verify the computation is correct', async () => {
+  it('should get and validate the result', async () => {
     task3 = await new Promise((resolve, reject) => {
       enigma.getTaskResult(task3)
         .on(eeConstants.GET_TASK_RESULT_RESULT, (result) => resolve(result))
         .on(eeConstants.ERROR, (error) => reject(error));
     });
+    const verifyTaskOutput = await enigma.verifyTaskOutput(task3);
+    const verifyTaskStatus = await enigma.verifyTaskStatus(task3);
+    expect(verifyTaskOutput).toEqual(true);
+    expect(verifyTaskStatus).toEqual(true);
     expect(task3.engStatus).toEqual('SUCCESS');
     expect(task3.encryptedAbiEncodedOutputs).toBeTruthy();
-    task3 = await enigma.decryptTaskResult(task3);
     expect(task3.usedGas).toBeTruthy();
     expect(task3.workerTaskSig).toBeTruthy();
-    expect(parseInt(task3.decryptedOutput, 16)).toEqual(12);
+    task3 = await enigma.decryptTaskResult(task3);
+    expect(task3.decryptedOutput).toEqual(utils.remove0x(addr2));
   });
 
 });
+
